provide option to create anchor links
Linking to procedure documentation is generally unwieldy, as the
default anchor id are generated as the word "procedure", followed by
entire procedure signature.

This commit introduces an option to additionally generate a simpler
anchor link for each definition, consisting of the prefix `def-`,
followed by the name of the defined binding.

Only Markdown mode supports this option, as there is no reliable way
of inserting raw HTML into svnwiki text.